Refactor duplicate publish call and add IPv6 support
If IPv6 is enabled on my host mdns tries to request an IPv6 address
(AAAA) first, this results in a ~3 seconds timeout before an IPv4
address is requested.

This patch adds the IPv6 address as well, to prevent four different
calls to publish() I've refactored it to take the cartesian product
of the hostnames and ip addresses and call publish in a for loop.

Signed-off-by: Dick Marinus <dick@mrns.nl>